requirements.txt
aiohttp==3.6.3
anyio==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==3.4.7
decorator==5.1.0
Deprecated==1.2.12
discord.py==1.7.3
distro==1.6.0
Flask==2.0.2
gcloud==0.18.3
googleapis-common-protos==1.53.0
h11==0.12.0
httpcore==0.13.7
httplib2==0.19.1
httpx==0.20.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.2
jwcrypto==0.9.1
lavalink==3.1.4
MarkupSafe==2.0.1
multidict==4.7.6
nextcord==2.0.0a2
oauth2client==4.1.3
pendulum==2.1.2
pipdeptree==2.1.0
pkce==1.0.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.5.0
python-dateutil==2.8.1
python-jwt==3.3.0
pytzdata==2020.1
ratelimit==2.2.1
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
spotipy==2.18.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uWSGI==2.0.20
validators==0.18.2
waitress==2.0.0
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.5.1
youtube-search-python==1.4.9
